Working static build time icon transformation
This will take any invocation of {{fa-icon 'coffee'}} or {{fa-icon
icon='coffee'}} and convert it into an SVG with other properties bound.
It only works for static icon names because it needs to lookup
information at build time from fontawesome and wouldn't be able to
re-computed those details at runtime if the icon name changed
dynamically.