Conditionally import all icon packs
There doesn't seem to be a way to lookup a module without requiring it
in node so I wrapped this in a try/catch and we just ignore errors when
an icon pack isn't loaded. This way users with pro icons don't have to
do any additional work to get those icons transformed.