Consolidate Configuration (#68)
In order to have a consistent configuration experience I've moved the
options that were in `ember-cli-build.js` into `environment.js`. Since
we need some configuration at both build and run time this is the best
place for it.

To provide for a smooth upgrade path existing configuration in
`ember-cli-build.js` will be respected, but a warning will be printed
during the build.  Any configuration in `environment.js` will supersede
anything in `ember-cli.build.js`